Log pageviews with piwik on all pages
